% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics.R
\name{add.alpha}
\alias{add.alpha}
\title{add.alpha}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Add an alpha value to a colour
}
\examples{
add.alpha(col, alpha)
}
\keyword{A}
