% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{CHECK1}
\alias{CHECK1}
\title{CHECK1}
\usage{
CHECK1(CLIB)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
function CHECK checks if CLIB is a perfect ordering of cliques
}
\examples{
CHECK1(CLIQUES)
}
\keyword{A}
