% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{CHECK1_OLD}
\alias{CHECK1_OLD}
\title{CHECK1_OLD}
\usage{
CHECK1_OLD(CLI)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
function CHECK checks if CLIB is a perfect ordering of cliques
}
\examples{
CHECK1_OLD(CLIQUES)
}
\keyword{A}
