% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{get.all.POCs}
\alias{get.all.POCs}
\title{get.all.POCs}
\usage{
get.all.POCs(CLIQUES, method)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
get.all.POCs descroption
}
\examples{
get.all.POCs(...)
}
\keyword{A}
