% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics (sergiusz wesołowski's
%   conflicted copy 2023-09-24).R, R/poc_functions_graphics.R
\name{Get.ALLPOCs.graph}
\alias{Get.ALLPOCs.graph}
\title{Get.ALLPOCs.graph}
\usage{
Get.ALLPOCs.graph(POClist, prefix = "", background = FALSE)

Get.ALLPOCs.graph(POClist, prefix = "", background = FALSE)
}
\arguments{
\item{POClist}{A list of cliques}

\item{prefix}{A prefix for the output file name}

\item{background}{Add background nodes to the graph}

\item{A}{Description of the parameter}
}
\description{
Create a graph of all possible orderings of cliques

Add an alpha value to a colour
}
\examples{
Get.ALLPOCs.graph(CLIQUES, prefix = 'ALLPOCs', background = TRUE)
Get.ALLPOCs.graph(CLIQUES, draw=TRUE)
}
\keyword{A}
