% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics (sergiusz wesołowski's
%   conflicted copy 2023-09-24).R, R/poc_functions_graphics.R
\name{Get.graph}
\alias{Get.graph}
\title{Get.graph}
\usage{
Get.graph(CLIQUES, draw = FALSE, prefix = "")

Get.graph(CLIQUES, draw = FALSE, prefix = "")
}
\arguments{
\item{CLIQUES}{A list of cliques.}

\item{draw}{If TRUE, generates a plot of the graph.}

\item{prefix}{A string that will be used to prefix the filename of the generated plot.}

\item{A}{Description of the parameter}
}
\value{
A graph object generated from the list of cliques.
}
\description{
Generate a graph from a list of cliques.

Add an alpha value to a colour
}
\examples{
Get.graph(CLIQUES)
Get.graph(CLIQUES, draw=TRUE)
}
\keyword{A}
