% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics (sergiusz wesołowski's
%   conflicted copy 2023-09-24).R, R/poc_functions_graphics.R
\name{Get.reduced.weighted.clique.graph}
\alias{Get.reduced.weighted.clique.graph}
\title{Get reduced weighted clique graph}
\usage{
Get.reduced.weighted.clique.graph(CLIQUES, draw = FALSE, prefix = "")

Get.reduced.weighted.clique.graph(CLIQUES, draw = FALSE, prefix = "")
}
\arguments{
\item{CLIQUES}{A list of cliques}

\item{draw}{Boolean indicating whether to draw the graph}

\item{prefix}{A prefix for the filename if saving the graph}

\item{A}{Description of the parameter}
}
\value{
The reduced weighted clique graph
}
\description{
Generates a reduced weighted clique graph based on input cliques

Add an alpha value to a colour
}
\examples{
Get.reduced.weighted.clique.graph(CLIQUES, draw=TRUE)
Get.reduced.weighted.clique.graph(CLIQUES, draw=TRUE)
}
\keyword{A}
