% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{get.single.POC}
\alias{get.single.POC}
\title{get.single.POC}
\usage{
get.single.POC(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
get.single.POC descroption
}
\examples{
get.single.POC(...)
}
\keyword{A}
