% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_MC.R
\name{get_CLI_next_length}
\alias{get_CLI_next_length}
\title{get_CLI_next_length}
\usage{
get_CLI_next_length(rmng_V, K)
}
\arguments{
\item{rmng_V}{A vector containing the vertices that have not yet been added to the graph.}

\item{K}{The size of the current clique.}
}
\value{
A numeric value indicating the length of the next clique to be added.
}
\description{
Generates a random length for the next clique to be added to a graph.
}
\details{
This function takes the remaining vertices and the size of the current clique and returns the length of the next clique to be added to the graph. The length of the next clique is chosen randomly, either K-1, K, or K+1, with equal probaability.
}
\examples{
get_CLI_next_length(c(1, 2, 3), 4)
}
