% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{is.POC}
\alias{is.POC}
\title{is.POC}
\usage{
is.POC(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
function CHECK checks if CLIB is a perfect ordering of cliques
}
\examples{
is.POC(CLIQUES)
}
\keyword{A}
