% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{is.removable}
\alias{is.removable}
\title{is.removalble}
\usage{
is.removable(path.r2s, A, W)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Check if removable
}
\examples{
is.removable(r,s,A,W )
}
\keyword{W}
