% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{lp2}
\alias{lp2}
\title{lp2}
\usage{
lp2(A)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Function lp2 draws (by rejection) randomly any non-empty proper subset of the set 1:n with the same probability.
}
\examples{
lp2(A)
}
\keyword{A}
