% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_MC.R
\name{markov_chain}
\alias{markov_chain}
\title{markov_chain}
\usage{
markov_chain(CLI_init, max_iter = 100, poc_shuffle = TRUE)
}
\arguments{
\item{CLI_init}{The initial graph, represented as a list of cliques.}

\item{max_iter}{The maximum number of iterations for the Markov chain. Default is 100.}
}
\value{
A list of graphs generated during the Markov chain.
}
\description{
Generates a Markov chain of graphs using the POC algorithm.
}
\details{
This function takes an initial graph and generates a Markov chain of graphs using the POC algorithm. The maximum number of iterations is specified by the max_iter argument. The output is a list of graphs generated during the Markov chain.
}
\examples{
markov_chain(list(c(1, 2), c(3, 4)))
}
