% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{matmod}
\alias{matmod}
\title{matmod}
\usage{
matmod(W, A)
}
\arguments{
\item{W}{The weighted coincidence matrix}

\item{A}{The set of vertices in R}
}
\description{
Puts zeros in the weighted incidence matrix A for those edges connecting vertices from R to vertices from R^c which are removable with respect to R
}
\examples{
matmod(W, A)
}
\keyword{A}
\keyword{W,}
