% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics (sergiusz wesołowski's
%   conflicted copy 2023-09-24).R, R/poc_functions_graphics.R
\name{prop_to_size}
\alias{prop_to_size}
\title{prop_to_size}
\usage{
prop_to_size(prop, mi = 0, ma = 1, power = 2)

prop_to_size(prop, mi = 0, ma = 1, power = 2)
}
\arguments{
\item{prop}{a numeric vector of properties to be resized}

\item{mi}{minimum size value}

\item{ma}{maximum size value}

\item{power}{the power to which the prop values should be raised}

\item{A}{Description of the parameter}
}
\value{
a vector of resized values based on the input parameters
}
\description{
Resize properties of a graph

Resize properties of a graph
}
\examples{
prop_to_size(c(1,2,3,4), 0, 10, 2)
prop_to_size()
}
\keyword{A}
\keyword{graph,}
\keyword{size}
