% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{rancli}
\alias{rancli}
\title{rancli}
\usage{
rancli(CLI)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
We draw a random permutation of cliques from CLI and accept it only if it is perfect
}
\examples{
rancli(CLIQUES)
}
\keyword{A}
