% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{rCLI}
\alias{rCLI}
\title{rCLI}
\usage{
rCLI(CLI)
}
\arguments{
\item{CLIQUES}{Description of the parameter}
}
\description{
We draw a random permutation of cliques from CLI and accept it only if it is perfect
}
\examples{
rancli(CLIQUES)
}
\keyword{A}
