% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{RCM}
\alias{RCM}
\title{RCM}
\usage{
RCM(W, o.k)
}
\arguments{
\item{W}{Description of the parameter}
}
\description{
THE CORE OF THE RCM WHICH PRINTS ALL PERFECT ORDERING OF "CLIQUES" IS FUNCTION: "RCM" THE FUNCTION "RCM.WRAPPER"	 JUST SETS THE STAGE
}
\examples{
RCM(...)
}
\keyword{W}
