% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{RCM.WRAPPER}
\alias{RCM.WRAPPER}
\title{RCM.WRAPPER}
\usage{
RCM.WRAPPER(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
THE CORE OF THE RCM WHICH PRINTS ALL PERFECT ORDERING OF "CLIQUES" IS FUNCTION: "RCM" THE FUNCTION "RCM.WRAPPER"	 JUST SETS THE STAGE
}
\examples{
RCM.WRAPPER(...)
}
\keyword{A}
