% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{RECURRENT.POC.SEARCH}
\alias{RECURRENT.POC.SEARCH}
\title{RECURRENT.POC.SEARCH}
\usage{
RECURRENT.POC.SEARCH(W, W.labels, o.k, CLIQUES, filename)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
THE CORE OF THE RECURSION WHICH PRINTS ALL PERFECT ORDERING OF "CLIQUES" IS FUNCTION: "RECURRENT.POC.SEARCH" THE FUNCTION "RECURRENT.POC.SEARCH.WRAPPER"	 JUST SETS THE STAGE
}
\examples{
RECURRENT.POC.SEARCH(...)
}
\keyword{A}
