% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{RECURSION.POC.SEARCH}
\alias{RECURSION.POC.SEARCH}
\title{RECURSION.POC.SEARCH}
\usage{
RECURSION.POC.SEARCH(W, W.labels, o.k)
}
\arguments{
\item{W}{Description of the parameter}
}
\description{
THE CORE OF THE RECURSION WHICH PRINTS ALL PERFECT ORDERING OF "CLIQUES" IS FUNCTION: "RECURSION.POC.SEARCH" THE FUNCTION "RECURSION.POC.SEARCH.WRAPPER"	 JUST SETS THE STAGE
}
\examples{
RECURSION.POC.SEARCH(...)
}
\keyword{W}
