% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{RECURSION.POC.SEARCH.WRAPPER}
\alias{RECURSION.POC.SEARCH.WRAPPER}
\title{RECURSION.POC.SEARCH.WRAPPER}
\usage{
RECURSION.POC.SEARCH.WRAPPER(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
THE CORE OF THE RECURSION WHICH PRINTS ALL PERFECT ORDERING OF "CLIQUES" IS FUNCTION: "RECURSION.POC.SEARCH" THE FUNCTION "RECURSION.POC.SEARCH.WRAPPER"	 JUST SETS THE STAGE
}
\examples{
RECURSION.POC.SEARCH.WRAPPER(...)
}
\keyword{A}
