% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{rPOC_recursion}
\alias{rPOC_recursion}
\title{rPOC}
\usage{
rPOC_recursion(n, CLIQUES, method = "recursion", replace = TRUE)
}
\arguments{
\item{n}{Description of the parameter}
}
\description{
rPOC description
}
\examples{
rPOC.unif(1,CLIQUES)
}
\keyword{random}
\keyword{sample}
