% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{rPOC_unif}
\alias{rPOC_unif}
\title{rPOC_unif}
\usage{
rPOC_unif(n, CLIQUES, method = "RCM", replace = TRUE, mh_burn = 100)

rPOC_unif(n, CLIQUES, method = "RCM", replace = TRUE, mh_burn = 100)
}
\arguments{
\item{n}{Number of POCs to be generated.}

\item{CLIQUES}{A list of cliques.}

\item{method}{Method to be used for sampling. Either 'RCM' or 'rejection'.}

\item{replace}{Boolean indicating whether to sample with or without replacement.}

\item{mh_burn}{Integer indicating the number of Metropolis-Hastings burn-in iterations to be used if method='M-H'.}
}
\value{
Returns a list of generated POCs.

Returns a list of generated POCs.
}
\description{
The function rPOC_unif generates random samples of perfect orderings of cliques (POCs) from the list of cliques using the rejection method.

The function rPOC_unif generates random samples of perfect orderings of cliques (POCs) from the list of cliques using the rejection method.
}
\examples{
rPOC_unif(100, CLIQUES)
rPOC_unif(100, CLIQUES)
}
