% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_MC.R
\name{shuffle_lists}
\alias{shuffle_lists}
\title{shuffle_lists}
\usage{
shuffle_lists(lists)
}
\arguments{
\item{lists}{A list of lists containing the objects to be merged and shuffled.}
}
\value{
A new list containing all objects from the input lists, shuffled but with the order of objects preserved
within each original list.
}
\description{
Merge and shuffle lists while preserving order of objects within each list
}
\details{
This function takes a list of lists and returns a new list containing all objects from the input lists,
with the order of the objects shuffled, but preserving the order of objects within each original list.
}
\examples{
list1 <- list("a", "b", "c")
list2 <- list("d", "e")
shuffled_lists(list(list1, list2))
}
