% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{single.POC}
\alias{single.POC}
\title{single.POC}
\usage{
single.POC(W, W.labels, o.k, CLIQUES, prob)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
single.POC descroption
}
\examples{
single.POC(...)
}
\keyword{A}
