% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{single.RCM}
\alias{single.RCM}
\title{single.RCM}
\usage{
single.RCM(W, o.k, prob)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
single.RCM descroption
}
\examples{
single.RCM(...)
}
\keyword{A}
