% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{single.RCM.WRAPPER}
\alias{single.RCM.WRAPPER}
\title{single.RCM.WRAPPER}
\usage{
single.RCM.WRAPPER(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
single.RCM.WRAPPER descroption
}
\examples{
single.RCM.WRAPPER(...)
}
\keyword{A}
