% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{split_into_3}
\alias{split_into_3}
\title{Function split_into_3}
\usage{
split_into_3(A)
}
\arguments{
\item{A}{A vector of elements to choose from}
}
\value{
A list of non-empty subsets of a set contaning all elements
}
\description{
Splits a set into 3 non-empty subsets..
}
\examples{
split_into_3(1:10)
}
\keyword{random}
\keyword{subset}
