% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{Wmat}
\alias{Wmat}
\title{Wmat}
\usage{
Wmat(CLI)
}
\arguments{
\item{CLI}{A list of cliques}
}
\description{
Constructs the weighted coincidence matrix for the set of cliques CLI with weight w(i,j) = number of vertices in C_i cap C_j
}
\examples{
Wmat(CLI)
}
\keyword{CLI}
