% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics (sergiusz wesołowski's
%   conflicted copy 2023-09-24).R, R/poc_functions_graphics.R
\name{add.alpha}
\alias{add.alpha}
\title{Add an alpha value to a colour}
\usage{
add.alpha(col, alpha = 1)

add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{A vector of colours.}

\item{alpha}{A value between 0 and 1 representing the level of transparency (default = 1).}

\item{A}{Description of the parameter}
}
\value{
A vector of colours with an alpha channel.
}
\description{
Add an alpha value to a colour vector using a given alpha value (default = 1).

Add an alpha value to a colour
}
\examples{
add.alpha(col, alpha)
}
\keyword{A}
