% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{get.all.POCs}
\alias{get.all.POCs}
\title{get.all.POCs}
\usage{
get.all.POCs(CLIQUES, method)
}
\arguments{
\item{CLIQUES}{a list of cliques}

\item{method}{the sampling method to use (either "RCM" or "rejection")}
}
\value{
a matrix of all perfect orderings of cliques
}
\description{
Gets all possible perfect orderings of cliques using RCM or rejection sampling method.
}
\examples{
get.all.POCs(list(1:3, 3:5, 2:4), "RCM")
}
\keyword{RCM,}
\keyword{cliques}
\keyword{rejection}
\keyword{sampling,}
