% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics.R
\name{Get.graph}
\alias{Get.graph}
\title{Get.graph}
\usage{
Get.graph(CLIQUES, draw = FALSE, prefix = "")
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Add an alpha value to a colour
}
\examples{
Get.graph(CLIQUES, draw=TRUE)
}
\keyword{A}
