% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{get.MH.poc}
\alias{get.MH.poc}
\title{get.MH.poc}
\usage{
get.MH.poc(CLIQUES, mh_burn = 100)
}
\arguments{
\item{n}{Description of the parameter}
}
\description{
get.MH.poc description
}
\examples{
get.MH.poc(CLIQUES,1000)
}
\keyword{random}
\keyword{sample}
