% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{is.decomposable}
\alias{is.decomposable}
\title{is.decomposable}
\usage{
is.decomposable(CLIQUES)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
get.single.POC descroption
}
\examples{
is.decomposable(CLIQUES)
}
\keyword{A}
