% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{is.decomposable}
\alias{is.decomposable}
\title{is.decomposable}
\usage{
is.decomposable(CLIQUES, method = "RCM")

is.decomposable(CLIQUES, method = "RCM")
}
\arguments{
\item{CLIQUES}{A list of cliques.}

\item{method}{Method to be used for generating POCs. Either 'RCM' or 'rejection'.}
}
\value{
Returns a boolean indicating if the list of cliques is decomposable.

Returns a boolean indicating if the list of cliques is decomposable.
}
\description{
The function is.decomposable determines if a list of cliques is decomposable by checking if a perfect ordering of cliques (POC) can be generated using the provided method.

The function is.decomposable determines if a list of cliques is decomposable by checking if a perfect ordering of cliques (POC) can be generated using the provided method.
}
\examples{
is.decomposable(CLIQUES, method='RCM')
is.decomposable(CLIQUES, method='RCM')
}
