% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{is.POC}
\alias{is.POC}
\title{is.POC}
\usage{
is.POC(CLIQUES)

is.POC(CLIQUES)
}
\arguments{
\item{CLIQUES}{List of cliques}
}
\description{
Checks if CLIQUES is a perfect ordering of cliques

Checks if CLIQUES is a perfect ordering of cliques
}
\examples{
is.POC(CLIQUES)
is.POC(CLIQUES)
}
\keyword{CLIQUES}
