% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{lp1}
\alias{lp1}
\title{Function lp1}
\usage{
lp1(A, p = 0.5)
}
\arguments{
\item{A}{A vector of elements to choose from}

\item{p}{The probability of including each element in the subset}
}
\value{
A randomly chosen non-empty subset of A
}
\description{
Draws (by rejection) randomly any non-empty subset of the set 1:n with the same probability.
}
\examples{
lp1(1:10)
}
\keyword{random}
\keyword{subset}
