% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{lp1}
\alias{lp1}
\title{lp1}
\usage{
lp1(A)
}
\arguments{
\item{A}{Description of the parameters}
}
\description{
Function lp1 draws (by rejection) randomly any non-empty subset of the set 1:n with the same probability.
}
\examples{
lp1()
}
\keyword{A}
