% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{lp2}
\alias{lp2}
\title{Function lp2}
\usage{
lp2(A, p = 0.5)

lp2(A, p = 0.5)
}
\arguments{
\item{A}{A vector of elements to choose from}

\item{p}{The probability of including each element in the subset}
}
\value{
A randomly chosen non-empty proper subset of A

A randomly chosen non-empty proper subset of A
}
\description{
Draws (by rejection) randomly any non-empty proper subset of the set 1:n with the same probability.

Draws (by rejection) randomly any non-empty proper subset of the set 1:n with the same probability.
}
\examples{
lp2(1:10)
lp2(1:10)
}
\keyword{random}
\keyword{subset}
