% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{lp3}
\alias{lp3}
\title{lp3}
\usage{
lp3(A)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Function lp3 draws (by rejection) randomly any subset of size >1 of the set 1:n with the same probability.
}
\examples{
lp3(A)
}
\keyword{A}
