% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{lp3}
\alias{lp3}
\title{Function lp3}
\usage{
lp3(A, p = 0.5)
}
\arguments{
\item{A}{A vector of elements to choose from}

\item{p}{The probability of including each element in the subset}
}
\value{
A randomly chosen subset of size >1 of A
}
\description{
Draws (by rejection) randomly any subset of size >1 of the set 1:n with the same probability.
}
\examples{
split_into_3(1:10)
}
\keyword{random}
\keyword{subset}
