% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{matmod}
\alias{matmod}
\title{matmod_new}
\usage{
matmod(W, A, CLI = CLI)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
FUNCTIONS Wmat AND matmod FORM THE TECHNICAL CORE OF THE ALGORITHM WHICH CONSTRUCTS A PERFECT PERMUTATION OF CLIQUES (ASSUMING THAT IT IS POSSIBLE). function matmod puts zeros in the weighted	incidence matrix A	for those edges connecting vertices from R to vertices from R^c which are removable with respect to R
}
\examples{
matmod(W,R,CLI)
}
\keyword{W}
