% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{poc1}
\alias{poc1}
\title{poc1}
\usage{
poc1(n)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
FUNCTION poc1 DRAWS (RANDOMLY, Non-uniformly) A PERFECT ORDERING OF CLIQUES OF n NODES	 (AND THUS A DECOMPOSABLE GRAPH) ITS OUTPUT CLI IS THE INPUT TO THE MAIN PROCEDURE
}
\examples{
poc1(10)
}
\keyword{A}
