% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions_graphics.R
\name{prop_to_size}
\alias{prop_to_size}
\title{prop_to_size}
\usage{
prop_to_size(prop, mi = 0, ma = 1, power = 2)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
Resize properties of a graph
}
\examples{
prop_to_size()
}
\keyword{A}
