% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{RCM}
\alias{RCM}
\title{RCM}
\usage{
RCM(W, o.k)

RCM(W, o.k)
}
\arguments{
\item{W}{The weighted coincidence matrix}

\item{o.k}{A vector of the indices of the ordered cliques in the perfect ordering}
}
\description{
Prints all perfect ordering of cliques

Prints all perfect ordering of cliques
}
\examples{
RCM(W, o.k)
RCM(W, o.k)
}
\keyword{W,}
\keyword{o.k}
