% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{RCM.WRAPPER}
\alias{RCM.WRAPPER}
\title{RCM.WRAPPER}
\usage{
RCM.WRAPPER(CLIQUES)

RCM.WRAPPER(CLIQUES)
}
\arguments{
\item{CLIQUES}{a list of cliques}
}
\value{
a matrix of all perfect orderings of cliques

a matrix of all perfect orderings of cliques
}
\description{
The core of the RCM which prints all perfect orderings of "cliques" is function: "RCM". The function "RCM.WRAPPER" sets the stage.

The core of the RCM which prints all perfect orderings of "cliques" is function: "RCM". The function "RCM.WRAPPER" sets the stage.
}
\examples{
RCM.WRAPPER(list(1:3, 3:5, 2:4))
RCM.WRAPPER(list(1:3, 3:5, 2:4))
}
\keyword{RCM,}
\keyword{function}
\keyword{wrapper}
