% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{rPOC_RCM}
\alias{rPOC_RCM}
\title{rPOC_RCM}
\usage{
rPOC_RCM(nsamples, CLIQUES, method = "RCM", replace = TRUE)

rPOC_RCM(nsamples, CLIQUES, method = "RCM", replace = TRUE)
}
\arguments{
\item{nsamples}{Number of POCs to be generated.}

\item{CLIQUES}{A list of cliques.}

\item{replace}{Boolean indicating whether to sample with or without replacement.}
}
\value{
Returns a list of generated POCs.

Returns a list of generated POCs.
}
\description{
The function rPOC_RCM generates random samples of perfect orderings of cliques (POCs) from the list of cliques using the RCM method.

The function rPOC_RCM generates random samples of perfect orderings of cliques (POCs) from the list of cliques using the RCM method.
}
\examples{
rPOC_RCM(100, CLIQUES)
rPOC_RCM(100, CLIQUES)
}
