% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{rPOC_unif}
\alias{rPOC_unif}
\title{rPOC_unif}
\usage{
rPOC_unif(n, CLIQUES, method = "RCM", replace = TRUE, mh_burn = 100)
}
\arguments{
\item{n}{Description of the parameter}
}
\description{
rPOC description
}
\examples{
rPOC.unif(1,CLIQUES)
}
\keyword{random}
\keyword{sample}
