% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{single.RCM}
\alias{single.RCM}
\title{single.RCM}
\usage{
single.RCM(W, o.k, prob)

single.RCM(W, o.k, prob)
}
\arguments{
\item{W}{a matrix of weights}

\item{o.k}{a vector of indices representing cliques already chosen}

\item{prob}{the probability of the perfect ordering}
}
\value{
a list with two items: a vector representing a perfect ordering of cliques and a probability

a list with two items: a vector representing a perfect ordering of cliques and a probability
}
\description{
The single RCM function uses recursive coalescence modeling to compute one perfect ordering of cliques.

The single RCM function uses recursive coalescence modeling to compute one perfect ordering of cliques.
}
\keyword{RCM,}
\keyword{cliques,}
\keyword{probability}
