% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions (sergiusz wesołowski's conflicted
%   copy 2023-09-24).R, R/poc_functions.R
\name{single.RCM.WRAPPER}
\alias{single.RCM.WRAPPER}
\title{single.RCM.WRAPPER}
\usage{
single.RCM.WRAPPER(CLIQUES)

single.RCM.WRAPPER(CLIQUES)
}
\arguments{
\item{CLIQUES}{A list of cliques.}
}
\value{
Returns a list containing the perfect ordering of cliques (POC) and the probability of the POC.

Returns a list containing the perfect ordering of cliques (POC) and the probability of the POC.
}
\description{
The function single.RCM.WRAPPER returns a perfect ordering of cliques (POC) of the list of cliques.

The function single.RCM.WRAPPER returns a perfect ordering of cliques (POC) of the list of cliques.
}
\examples{
single.RCM.WRAPPER(CLIQUES)
single.RCM.WRAPPER(CLIQUES)
}
