% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_functions.R
\name{Wmat}
\alias{Wmat}
\title{Wmat}
\usage{
Wmat(CLI)
}
\arguments{
\item{A}{Description of the parameter}
}
\description{
FUNCTIONS Wmat AND matmod FORM THE TECHNICAL CORE OF THE ALGORITHM WHICH CONSTRUCTS A PERFECT PERMUTATION OF CLIQUES (ASSUMING THAT IT IS POSSIBLE). function Wmat constructs the weighted coincidence matrix for the set	of cliques CLI with weight w(i,j) = number of vertices in C_i cap C_j
}
\examples{
Wmat(CLIQUES)
}
\keyword{A}
